% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha2theta.R
\name{alpha2theta}
\alias{alpha2theta}
\title{Transforms unconstrained parameters to constrainted  parameters}
\usage{
alpha2theta(family, param)
}
\arguments{
\item{family}{distribution name; run the command distributions() for help}

\item{param}{unconstrained parameters of the univariate distribution}
}
\value{
\item{theta}{constrained parameters}
}
\description{
This function computes the constrained  parameters theta of a univariate distribution from the unconstrainted parameter alpha.
}
\keyword{internal}
