% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfun.R
\name{bfun}
\alias{bfun}
\title{Bootstrap for the univariate distributions}
\usage{
bfun(theta, Q, ZI, family, start, n, size, max_iter, eps)
}
\arguments{
\item{theta}{parameters}

\item{Q}{transition matrix for the regimes}

\item{ZI}{1 if zero-inflated, 0 otherwise (default)}

\item{family}{distribution name; (run the command distributions() for help)}

\item{n}{number of simulated observations}

\item{size}{additional parameter for some discrete distributions; run the command distributions() for help}

\item{max_iter}{maximum number of iterations of the EM algorithm; suggestion 10000}

\item{eps}{precision (stopping criteria); suggestion 0.0001}
}
\value{
Internal function used for the parametric bootstrap
}
\description{
Bootstrap for the univariate distributions
}
\keyword{internal}
