% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfun.R
\name{bootstrapfun}
\alias{bootstrapfun}
\title{Bootstrap for the univariate distributions}
\usage{
bootstrapfun(
  Q,
  family,
  theta,
  n,
  size = 0,
  max_iter = 10000,
  eps = 0.001,
  useFest = 1
)
}
\arguments{
\item{Q}{transition matrix;}

\item{family}{distribution name; (run the command distributions() for help)}

\item{theta}{parameters}

\item{n}{number of simulated observations}

\item{size}{additional parameter for some discrete distributions; run the command distributions() for help}

\item{max_iter}{maximum number of iterations of the EM algorithm; suggestion 10000}

\item{eps}{precision (stopping criteria); suggestion 0.001}

\item{n_sample}{number of bootstrap samples; suggestion 1000}

\item{UseFest}{1 (default) to use the first estimated parameters as starting value for the bootstrap, 0 otherwise}
}
\value{
Internal function used for the parametric bootstrap
}
\description{
Bootstrap for the univariate distributions
}
\keyword{internal}
