% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDF.R
\name{CDF}
\alias{CDF}
\title{Cumulative distribution function}
\usage{
CDF(family, y, param, size = 0)
}
\arguments{
\item{family}{distribution name; run the function distributions() for help}

\item{y}{observations}

\item{param}{parameters of the distribution; (1 x p)}

\item{size}{additional parameter for some discrete distributions; run the command distributions() for help}
}
\value{
\item{f}{cdf}
}
\description{
This function computes the cumulative distribution function (cdf) of a univariate distribution
}
