% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDF_est.R
\name{CDF_est}
\alias{CDF_est}
\title{Cumulative distribution function}
\usage{
CDF_est(family, y, param, size = 0, u_Ros = 0)
}
\arguments{
\item{family}{distribution name; run the command distributions() for help}

\item{y}{observations}

\item{param}{parameters of the distribution}

\item{size}{additional parameter for some discrete distributions}

\item{u_Ros}{used internally for the computation of the Rosenblatt transform for non continuous distributions}
}
\value{
\item{f}{cdf}
}
\description{
This function computes the cumulative distribution function of an univariate distribution (used) in the CVM calculation
}
\keyword{internal}
