% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{distributions}
\alias{distributions}
\title{The names and descriptions of the univariate distributions}
\usage{
distributions()
}
\value{
No returned value, allows the users to know the different distributions and parameters
}
\description{
This function allows the users to find the details on the available distributions.
}
