% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphEstim.R
\name{graphEstim}
\alias{graphEstim}
\title{Graphs}
\usage{
graphEstim(y, reg, theta, family, pred_l, pred_e)
}
\arguments{
\item{y}{observations}

\item{reg}{number of regimes}

\item{theta}{estimated parameters; (r x p)}

\item{family}{distribution name; run the function distributions() for help}

\item{pred_l}{estimated regime using lambda}

\item{pred_e}{estimated regime using eta}
}
\value{
No returned value; produces figures of interest for the HMM model
}
\description{
This function shows the graphs resulting from the estimation of a HMM model
}
