% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDF_unc.R
\name{PDF_unc}
\alias{PDF_unc}
\title{Probability density function}
\usage{
PDF_unc(family, y, param, size = 0)
}
\arguments{
\item{family}{distribution name; run the command distributions() for help}

\item{y}{observations}

\item{param}{parameters of the distribution (unconstrained,  in R)}

\item{size}{additional parameter for some discrete distributions}
}
\value{
\item{f}{pdf}
}
\description{
This function computes the probability density function of a univariate distribution with unconstrained parameters
}
\keyword{internal}
