% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snd1.R
\name{Snd1}
\alias{Snd1}
\title{Cramer-von Mises statistic for the goodness-of-fit test of the null hypothesis of a univariate uniform distribution over [0,1]}
\usage{
Snd1(U)
}
\arguments{
\item{U}{vector of pseudos-observations (approximating uniform)}
}
\value{
\item{sta}{Cramer-von Mises statistic}
}
\description{
This function computes the Cramer-von Mises statistic Sn for goodness-of-fit of the null hypothesis of a univariate uniform distrubtion over [0,1]
}
