% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCM.R
\name{TCM}
\alias{TCM}
\title{Tail conditional median function}
\usage{
TCM(family, p, param, size = 0, Nsim = 25000)
}
\arguments{
\item{family}{distribution name; run the command distributions() for help}

\item{p}{(1  x 1) values  between 0 and 1}

\item{param}{parameters of the distribution}

\item{size}{additional parameter for some discrete distributions}

\item{Nsim}{number of simulations}
}
\value{
\item{tcm}{tail conditional median}
}
\description{
This function computes the tail conditional median of a univariate distribution
}
\examples{
family = "gaussian"

Q = 1 ; theta = matrix(c(-1.5, 1.7),1,2) ;
tcm = TCM(family, (0.01), theta)
print('Tail conditional mean : ')
print(tcm$tcm)


}
