% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDF.R
\name{PDF}
\alias{PDF}
\title{Probability density function}
\usage{
PDF(family, y, param)
}
\arguments{
\item{family}{distribution name; run the function distributions() for help}

\item{y}{observations}

\item{param}{parameters of the distribution; (1 x p)}
}
\value{
\item{f}{pdf}
}
\description{
This function computes the probability density function (pdf) of a univariate distribution
}
