% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta2alpha.R
\name{theta2alpha}
\alias{theta2alpha}
\title{Transform constrained parameters  to unconstrained parameters}
\usage{
theta2alpha(family, param)
}
\arguments{
\item{family}{distribution name; run the command distributions() for help}

\item{param}{constrained parameters of the univariate distribution}
}
\value{
\item{alpha}{constrained parameters}
}
\description{
This function computes the unconstrained parameters alpha of a univariate distribution corresponding to constrainted parameters theta.
}
\keyword{internal}
